Support additional response types
This uses the Symfony base request rather than the Laravel one. This will allow more response types to be serialized. This also fixes an issue when trying to serialize a response that has an exception set on it. Some exceptions, for example the ValidationException, contain closures. When trying to serialize these exceptions they fail since closures cannot be serialized.